[game] cleaner CLOS usage
I'm getting better at specifying interfaces. Rules of thumb:

1. `:accessor` only needs to be specified once, and can be done on the
parent class that initially defines the slot.

2. Parent classes are only useful if you have at least one `defmethod`
that actually dispatches on the parent. It's when you want default
behaviour that is truly independent of the children themselves.
Otherwise, it's sufficient to avoid a common parent and just have slots
of the same name.

3. Avoid deep inheritance trees. So while you could argue that any `bug`
needs to also be a `facing-entity` to be useful, inherit `facing-entity`
on the concrete child class (bug-fly), never on the abstract
parent (bug). The abstract parent would only need to inherit something
itself if in one of its `defmethod`s you needed to call a certain
function from the grandparent, and there's no guarantee that the child
inherited the grandparent for you.

4. Avoid `:initform` on the parent unless there is truly a default that
applies logically to all children. Otherwise, enforce each child to be
proactive about defining its defaults. We'd rather have Conditions
thrown proactively than have strange defaults piped in through parents
that are hard to track down.